05/12/17 9:27

+++ app/src/main/java/com/example/owner/android5778_3965_2493_00/controller/addCarActivity.java
@@ -59,8 +59,8 @@ public class addCarActivity extends Activity implements View.OnClickListener {
         final ContentValues contentValues = new ContentValues();
         try {
             contentValues.put(RentConst.CarConst.CARNUMBER, this.CarNumberEditText.getText().toString());
-            contentValues.put(RentConst.CarConst.HOUSEBRANCH, this.HouseBranchEditText.getText().toString());
-            contentValues.put(RentConst.CarConst.MILEAGE, this.MileAgeEditText.getText().toString());
+            contentValues.put(RentConst.CarConst.HOUSEBRANCH, Integer.valueOf(this.HouseBranchEditText.getText().toString()));
+            contentValues.put(RentConst.CarConst.MILEAGE, Float.valueOf(this.MileAgeEditText.getText().toString()));
             contentValues.put(RentConst.CarConst.MODELNAME, this.ModelNameEditText.getText().toString());
             //DBManagerFactory.getManager().addCar(contentValues);
         }
